# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Tree API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/trees/{tree_id}/invites": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "List Invites",
        "operationId": "list_invites_api_trees__tree_id__invites_get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteOut"
                  },
                  "title": "Response List Invites Api Trees  Tree Id  Invites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Create Invite",
        "operationId": "create_invite_api_trees__tree_id__invites_post",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/invites/{invite_id}": {
      "delete": {
        "tags": [
          "invites"
        ],
        "summary": "Revoke Invite",
        "operationId": "revoke_invite_api_trees__tree_id__invites__invite_id__delete",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{token}": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "Preview Invite",
        "operationId": "preview_invite_api_invites__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{token}/accept": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Accept Invite",
        "operationId": "accept_invite_api_invites__token__accept_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InviteAccepted": {
        "properties": {
          "tree_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tree Id"
          }
        },
        "type": "object",
        "required": [
          "tree_id"
        ],
        "title": "InviteAccepted"
      },
      "InviteCreate": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "contributor",
              "personal"
            ],
            "title": "Role"
          },
          "subtree_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtree Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "InviteCreate"
      },
      "InviteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "subtree_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtree Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "role",
          "subtree_id",
          "email",
          "person_id",
          "expires_at",
          "accepted_at",
          "revoked"
        ],
        "title": "InviteOut"
      },
      "InvitePreview": {
        "properties": {
          "tree_name": {
            "type": "string",
            "title": "Tree Name"
          },
          "subtree_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtree Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "person_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Name"
          },
          "usable": {
            "type": "boolean",
            "title": "Usable"
          }
        },
        "type": "object",
        "required": [
          "tree_name",
          "subtree_name",
          "role",
          "person_name",
          "usable"
        ],
        "title": "InvitePreview"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "personal",
          "contributor",
          "admin",
          "owner"
        ],
        "title": "Role",
        "description": "Membership roles, lowest to highest privilege."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}